Add files via upload
Proposal
Before you dive into building out your final capstone, we’d like you to craft a
proposal for your project. This should allow you to clarify what you’re
looking to build, as well as get feedback on that objective before committing
the full time it would take to build it.
The proposal should be approximately one page long and answer the
following questions:
o What is the problem you are attempting to solve?
We are trying to determine if the customer will buy the bought item again or
not and if we can recommend the customer any product that matches
his/her choice of buying product history. This is a good model since it will
make easier for the selling company to predict if the customer will reorder
the bought item, which will help company to keep the inventory of the item
up to date. This will benefit company in getting more profits.

o How is your solution valuable?
This solution is most valuable for this case study since this deep down analysis will
impact the organization in profit or loss. We will analyze the data in every
dimension which will benefit the organization and will be beneficial for the
customer as well. Researches have shown
that if you analyze the data of the organization and try to identify the trends, it
helps your business model to improve the process and maximize your profit. This
solution uses the accuracy of the machine learning and trusts on the technology
to predict if the customer will reorder the item or not.
o What is your data source and how will you access it?

I have the downloaded the data from kaggle website. The URL of the data source
is https://www.kaggle.com/c/instacart-market-basket-analysis/data.We will read
the downloaded files using python functions.

o What techniques from the course do you anticipate using?
I will be using Supervised Learning, Unsupervised learning and Deep learning to
predict the outcome. I will determine which model will be preferred based on the
accuracy scored by each model. I usually prefer supervised learning since it gives a
better accuracy over unsupervised. I will drill down the data to understand the
data and will analyze the customer behavior to predict their future behavior using
visualization python tools such as Matplotlib, seaplot, ggplot etc.

o What do you anticipate to be the biggest challenge you’ll face?
The most common challenge will be to clean the data properly and remove all the
garbage values or null values. This impacts the data since the tool/technology is
not able to understand the data correctly and we get unacceptable results. The
biggest challenge would be to get the machine learning models have better
accuracy so that the prediction is better and more trustable. This can only be
done by doing more analysis and derive more useful features so that machine
learning models are able to analyze the data in a better way.